+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Improvements
 
 * improved logging for forker and forked child processes
+* test variables sorted by name for prettier export diffs
 
 ## 5.0.5 (2019-11-02)